Use base macro for lookup-reply-resolve macros
This adds `nephthys.macros.types.ReplyMacro`, which is a simple base
macro class for macros like `?shipwrights` and `?identity` that follow
the lookup-reply-resolve logic.

Closes #182
